requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
curlify==3.0.0
facebook_business==22.0.5
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
pycountry==24.6.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
